[Ide] Split document map and project data into own files and classes
Makes it easy to reason about document and project data handling, and
makes it testable
